+++ plugins/DamageBreakdownChart/Views/BreakdownView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace DamageBreakdownChart.Views;
+
+public partial class BreakdownView : UserControl
+{
+    public BreakdownView()
+    {
+        InitializeComponent();
+    }
+}