+++ plugins/DamageBreakdownChart/Views/TreemapView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace DamageBreakdownChart.Views;
+
+public partial class TreemapView : UserControl
+{
+    public TreemapView()
+    {
+        InitializeComponent();
+    }
+}